db: move settings into separate file
